+++ src/lib.rs
@@ -1,8 +1,8 @@
 //! **In-Memory Data Storage**
-//! 
+//!
 //! Project Stage
-//! 
-//! **Research:** This project is at the design stage, with some sketches of work but nothing usable yet.
+//!
+//! **Development**: this project already has milestone releases, but is still under active development, you should not expect full stability yet.
 
 pub mod storage;
 pub mod utils;