+++ src/utils/mod.rs
@@ -1,1 +1,2 @@
+pub mod build_profile;
 pub mod config;